fix(research): validate bounded note marker labels
